+++ imml/classify/_muse/code_encoder.py
@@ -10,10 +10,10 @@ except ImportError:
     deepmodule_installed = False
     deepmodule_error = "Module 'deep' needs to be installed. See https://imml.readthedocs.io/stable/main/installation.html#optional-dependencies"
 
-nn.Module = nn.Module if deepmodule_installed else object
+Module = nn.Module if deepmodule_installed else object
 
 
-class RelTemporalEncoding(nn.Module):
+class RelTemporalEncoding(Module):
     def __init__(self, embedding_size, max_len=5000):
         super(RelTemporalEncoding, self).__init__()
         position = torch.arange(0., max_len).unsqueeze(1)
@@ -29,7 +29,7 @@ class RelTemporalEncoding(nn.Module):
         return timestamps_emb
 
 
-class Attention(nn.Module):
+class Attention(Module):
     def forward(self, query, key, value, mask, dropout=None):
         scores = torch.matmul(query, key.transpose(-2, -1)) / math.sqrt(query.size(-1))
         if mask is not None:
@@ -41,7 +41,7 @@ class Attention(nn.Module):
         return torch.matmul(p_attn, value), p_attn
 
 
-class MultiHeadedAttention(nn.Module):
+class MultiHeadedAttention(Module):
     def __init__(self, h, d_model, dropout=0.1):
         super(MultiHeadedAttention, self).__init__()
         assert d_model % h == 0
@@ -78,7 +78,7 @@ class MultiHeadedAttention(nn.Module):
         return self.output_linear(x)
 
 
-class PositionwiseFeedForward(nn.Module):
+class PositionwiseFeedForward(Module):
     def __init__(self, d_model, d_ff, dropout=0.1):
         super(PositionwiseFeedForward, self).__init__()
         self.w_1 = nn.Linear(d_model, d_ff)
@@ -93,7 +93,7 @@ class PositionwiseFeedForward(nn.Module):
         return x
 
 
-class SublayerConnection(nn.Module):
+class SublayerConnection(Module):
     """
     A residual connection followed by a layer norm.
     Note for code simplicity the norm is first as opposed to last.
@@ -109,7 +109,7 @@ class SublayerConnection(nn.Module):
         return x + self.dropout(sublayer(self.norm(x)))
 
 
-class TransformerBlock(nn.Module):
+class TransformerBlock(Module):
     """
     Transformer Block = MultiHead Attention + Feed Forward with sublayer connection
     """
@@ -140,7 +140,7 @@ class TransformerBlock(nn.Module):
         return x
 
 
-class CodeEncoder(nn.Module):
+class CodeEncoder(Module):
     def __init__(
             self,
             tokenizer,