+++ imml/classify/_muse/ffn_encoder.py
@@ -8,10 +8,10 @@ except ImportError:
     deepmodule_installed = False
     deepmodule_error = "Module 'deep' needs to be installed. See https://imml.readthedocs.io/stable/main/installation.html#optional-dependencies"
 
-nn.Module = nn.Module if deepmodule_installed else object
+Module = nn.Module if deepmodule_installed else object
 
 
-class FFNEncoder(nn.Module):
+class FFNEncoder(Module):
     def __init__(self, input_dim, hidden_dim, output_dim, num_layers, dropout_prob=0.5, device="cpu"):
         super(FFNEncoder, self).__init__()