+++ imml/classify/_muse/gnn.py
@@ -10,7 +10,7 @@ except ImportError:
     deepmodule_installed = False
     deepmodule_error = "Module 'deep' needs to be installed. See https://imml.readthedocs.io/stable/main/installation.html#optional-dependencies"
 
-nn.Module = nn.Module if deepmodule_installed else object
+Module = nn.Module if deepmodule_installed else object
 MessagePassing = MessagePassing if deepmodule_installed else object
 
 
@@ -46,7 +46,7 @@ class EdgeSAGEConv(MessagePassing):
                                                      self.edge_channels)
 
 
-class GNNStack(nn.Module):
+class GNNStack(Module):
     def __init__(self, node_channels, edge_channels, normalize_embs, num_layers, dropout):
         super(GNNStack, self).__init__()
         self.node_channels = node_channels
@@ -90,7 +90,7 @@ class GNNStack(nn.Module):
         return x
 
 
-class MML(nn.Module):
+class MML(Module):
     def __init__(self, num_modalities, hidden_channels, normalize_embs, num_layers, dropout, output_dim):
         super(MML, self).__init__()
         self.output_dim = output_dim
@@ -165,7 +165,7 @@ class MML(nn.Module):
             loss = F.cross_entropy(l, y)
         return loss
 
-    def forward(self, Xs, observed_mod_indicator, y, y_indicator):
+    def forward(self, Xs, missing_mod_indicator, y, y_indicator):
         batch_size = Xs[0].size(0)
         hidden_dim = Xs[0].size(1)
         x = torch.stack(Xs, dim=1)
@@ -173,13 +173,13 @@ class MML(nn.Module):
         g_patient_nodes = torch.ones(batch_size, hidden_dim)
         g_patient_nodes = g_patient_nodes.to(x.device)
         g_nodes = torch.cat([g_patient_nodes, self.modality_nodes], dim=0)
-        g_edge_index = observed_mod_indicator.nonzero().t()
+        g_edge_index = missing_mod_indicator.nonzero().t()
         g_edge_index[1] += batch_size
         g_edge_index = torch.cat([g_edge_index, g_edge_index.flip([0])], dim=1)
-        g_edge_attr = x[observed_mod_indicator].repeat(2, 1)
+        g_edge_attr = x[missing_mod_indicator].repeat(2, 1)
         z = self.gnn(g_nodes, g_edge_attr, g_edge_index)
 
-        ag_x_flag = observed_mod_indicator.clone()
+        ag_x_flag = missing_mod_indicator.clone()
         ag_x_flag = self.edgedrop(ag_x_flag)
         ag_patient_nodes = torch.ones(batch_size, hidden_dim)
         ag_patient_nodes = ag_patient_nodes.to(x.device)
@@ -215,7 +215,7 @@ class MML(nn.Module):
 
         return 0.5 * unsup_loss + 0.5 * sup_loss + cls_loss
 
-    def inference(self, Xs, observed_mod_indicator):
+    def inference(self, Xs, missing_mod_indicator):
         batch_size = Xs[0].size(0)
         hidden_dim = Xs[0].size(1)
         x = torch.stack(Xs, dim=1)
@@ -223,10 +223,10 @@ class MML(nn.Module):
         g_patient_nodes = torch.ones(batch_size, hidden_dim)
         g_patient_nodes = g_patient_nodes.to(x.device)
         g_nodes = torch.cat([g_patient_nodes, self.modality_nodes], dim=0)
-        g_edge_index = observed_mod_indicator.nonzero().t()
+        g_edge_index = missing_mod_indicator.nonzero().t()
         g_edge_index[1] += batch_size
         g_edge_index = torch.cat([g_edge_index, g_edge_index.flip([0])], dim=1)
-        g_edge_attr = x[observed_mod_indicator].repeat(2, 1)
+        g_edge_attr = x[missing_mod_indicator].repeat(2, 1)
 
         z = self.gnn(g_nodes, g_edge_attr, g_edge_index)
         z = z[:batch_size]