+++ imml/load/integrao_dataset.py
@@ -13,10 +13,10 @@ except ImportError:
     deepmodule_installed = False
     deepmodule_error = "Module 'deep' needs to be installed. See https://imml.readthedocs.io/stable/main/installation.html#optional-dependencies"
 
-torch.utils.data.Dataset = torch.utils.data.Dataset if deepmodule_installed else object
+Dataset = torch.utils.data.Dataset if deepmodule_installed else object
 
 
-class IntegrAODataset(torch.utils.data.Dataset):
+class IntegrAODataset(Dataset):
     r"""
     This class provides a `torch.utils.data.Dataset` implementation for handling multi-modal datasets with `IntegrAO`.