+++ tests/load/test_m3care_dataset.py
@@ -13,24 +13,23 @@ def sample_data():
     n_mods = 3
     Xs = [torch.rand((n_samples, 10)) for _ in range(n_mods)]
     y = torch.randint(0, 2, (n_samples,), dtype=torch.float)
-    observed_mod_indicator = torch.ones((n_samples, n_mods), dtype=torch.bool)
-    return Xs, y, observed_mod_indicator
+    return Xs, y
 
 
 def test_deepmodule_not_installed(sample_data):
-    Xs, y, observed_mod_indicator = sample_data
-    M3CareDataset(Xs=Xs, y=y, observed_mod_indicator=observed_mod_indicator)
+    Xs, y = sample_data
+    M3CareDataset(Xs=Xs, y=y)
     with patch.dict(sys.modules, {"torch": None}):
         import imml.load.m3care_dataset as module_mock
         importlib.reload(module_mock)
         with pytest.raises(ImportError, match="Module 'deep' needs to be installed."):
-            M3CareDataset(Xs=Xs, y=y, observed_mod_indicator=observed_mod_indicator)
+            M3CareDataset(Xs=Xs, y=y)
     importlib.reload(module_mock)
 
 
 def test_default_params(sample_data):
-    Xs, y, observed_mod_indicator = sample_data
-    dataset = M3CareDataset(Xs=Xs, y=y, observed_mod_indicator=observed_mod_indicator)
+    Xs, y = sample_data
+    dataset = M3CareDataset(Xs=Xs, y=y)
     assert len(dataset) == len(y)
     assert hasattr(dataset, 'Xs')
     assert hasattr(dataset, 'y')
@@ -47,33 +46,23 @@ def test_invalid_params():
     n_samples = 5
     Xs = [torch.rand((n_samples, 10)) for _ in range(3)]
     y = torch.randint(0, 2, (n_samples,), dtype=torch.float)
-    observed_mod_indicator = torch.ones((n_samples, 3), dtype=torch.bool)
     with pytest.raises(ValueError, match="Invalid Xs."):
-        M3CareDataset(Xs="not_a_list", y=y, observed_mod_indicator=observed_mod_indicator)
+        M3CareDataset(Xs="not_a_list", y=y)
     with pytest.raises(ValueError, match="Invalid Xs."):
-        M3CareDataset(Xs=[], y=y, observed_mod_indicator=observed_mod_indicator)
+        M3CareDataset(Xs=[], y=y)
     with pytest.raises(ValueError, match="Invalid Xs."):
-        M3CareDataset(Xs=[torch.rand((n_samples, 10)), torch.rand((0, 10))], y=y,
-                     observed_mod_indicator=observed_mod_indicator)
+        M3CareDataset(Xs=[torch.rand((n_samples, 10)), torch.rand((0, 10))], y=y)
     with pytest.raises(ValueError, match="Invalid Xs."):
-        M3CareDataset(Xs=[torch.rand((n_samples, 10)), torch.rand((n_samples+1, 10))], y=y,
-                     observed_mod_indicator=observed_mod_indicator)
+        M3CareDataset(Xs=[torch.rand((n_samples, 10)), torch.rand((n_samples+1, 10))], y=y)
     with pytest.raises(ValueError, match="Invalid y."):
-        M3CareDataset(Xs=Xs, y=None, observed_mod_indicator=observed_mod_indicator)
+        M3CareDataset(Xs=Xs, y=None)
     with pytest.raises(ValueError, match="Invalid y."):
-        M3CareDataset(Xs=Xs, y=torch.randint(0, 2, (n_samples+1,)),
-                     observed_mod_indicator=observed_mod_indicator)
-    with pytest.raises(ValueError, match="Invalid observed_mod_indicator."):
-        M3CareDataset(Xs=Xs, y=y, observed_mod_indicator=None)
-    with pytest.raises(ValueError, match="Invalid observed_mod_indicator."):
-        M3CareDataset(Xs=Xs, y=y, observed_mod_indicator=torch.ones((n_samples+1, 3)))
-    with pytest.raises(ValueError, match="Invalid observed_mod_indicator."):
-        M3CareDataset(Xs=Xs, y=y, observed_mod_indicator=torch.ones((n_samples, 2)))
+        M3CareDataset(Xs=Xs, y=torch.randint(0, 2, (n_samples+1,)))
 
 
 def test_getitem(sample_data):
-    Xs, y, observed_mod_indicator = sample_data
-    dataset = M3CareDataset(Xs=Xs, y=y, observed_mod_indicator=observed_mod_indicator)
+    Xs, y = sample_data
+    dataset = M3CareDataset(Xs=Xs, y=y)
     for i in range(len(dataset)):
         sample = dataset[i]
         assert isinstance(sample, tuple)
@@ -88,15 +77,14 @@ def test_getitem(sample_data):
         assert y_idx.item() == y[i].item()
         observed_mod_indicator_idx = sample[2]
         assert isinstance(observed_mod_indicator_idx, torch.Tensor)
-        assert torch.all(observed_mod_indicator_idx == observed_mod_indicator[i])
 
 
 def test_missing_values(sample_data):
-    Xs, y, observed_mod_indicator = sample_data
-    observed_mod_indicator[0, 0] = False
-    observed_mod_indicator[1, 1] = False
-    observed_mod_indicator[2, 2] = False
-    dataset = M3CareDataset(Xs=Xs, y=y, observed_mod_indicator=observed_mod_indicator)
+    Xs, y = sample_data
+    Xs[0][0, :] = torch.nan
+    Xs[1][1, :] = torch.nan
+    Xs[2][2, :] = torch.nan
+    dataset = M3CareDataset(Xs=Xs, y=y)
     sample = dataset[0]
     assert not sample[2][0].item()
     assert sample[2][1].item()