+++ tests/load/test_muse_dataset.py
@@ -14,29 +14,27 @@ def sample_data():
     n_mods = 3
     Xs = [torch.rand((n_samples, 10)) for _ in range(n_mods)]
     y = torch.randint(0, 2, (n_samples,), dtype=torch.float)
-    observed_mod_indicator = torch.ones((n_samples, n_mods), dtype=torch.bool)
-    y_indicator = torch.ones((n_samples,), dtype=torch.bool)
-    return Xs, y, observed_mod_indicator, y_indicator
+    return Xs, y
 
 
 def test_deepmodule_not_installed(sample_data):
-    Xs, y, observed_mod_indicator, y_indicator = sample_data
-    MUSEDataset(Xs=Xs, y=y, observed_mod_indicator=observed_mod_indicator, y_indicator=y_indicator)
+    Xs, y = sample_data
+    MUSEDataset(Xs=Xs, y=y)
     with patch.dict(sys.modules, {"torch": None}):
         import imml.load.muse_dataset as module_mock
         importlib.reload(module_mock)
         with pytest.raises(ImportError, match="Module 'deep' needs to be installed."):
-            MUSEDataset(Xs=Xs, y=y, observed_mod_indicator=observed_mod_indicator, y_indicator=y_indicator)
+            MUSEDataset(Xs=Xs, y=y)
     importlib.reload(module_mock)
 
 
 def test_default_params(sample_data):
-    Xs, y, observed_mod_indicator, y_indicator = sample_data
-    dataset = MUSEDataset(Xs=Xs, y=y, observed_mod_indicator=observed_mod_indicator, y_indicator=y_indicator)
+    Xs, y = sample_data
+    dataset = MUSEDataset(Xs=Xs, y=y)
     assert len(dataset) == len(y)
     assert hasattr(dataset, 'Xs')
     assert hasattr(dataset, 'y')
-    assert hasattr(dataset, 'observed_mod_indicator')
+    assert hasattr(dataset, 'missing_mod_indicator')
     assert hasattr(dataset, 'y_indicator')
     sample = dataset[0]
     assert isinstance(sample, tuple)
@@ -51,39 +49,23 @@ def test_invalid_params():
     n_samples = 5
     Xs = [torch.rand((n_samples, 10)) for _ in range(3)]
     y = torch.randint(0, 2, (n_samples,), dtype=torch.float)
-    observed_mod_indicator = torch.ones((n_samples, 3), dtype=torch.bool)
-    y_indicator = torch.ones((n_samples,), dtype=torch.bool)
     with pytest.raises(ValueError, match="Invalid Xs."):
-        MUSEDataset(Xs="not_a_list", y=y, observed_mod_indicator=observed_mod_indicator, y_indicator=y_indicator)
+        MUSEDataset(Xs="not_a_list", y=y)
     with pytest.raises(ValueError, match="Invalid Xs."):
-        MUSEDataset(Xs=[], y=y, observed_mod_indicator=observed_mod_indicator, y_indicator=y_indicator)
+        MUSEDataset(Xs=[], y=y)
     with pytest.raises(ValueError, match="Invalid Xs."):
-        MUSEDataset(Xs=[torch.rand((n_samples, 10)), torch.rand((0, 10))], y=y,
-                   observed_mod_indicator=observed_mod_indicator, y_indicator=y_indicator)
+        MUSEDataset(Xs=[torch.rand((n_samples, 10)), torch.rand((0, 10))], y=y)
     with pytest.raises(ValueError, match="Invalid Xs."):
-        MUSEDataset(Xs=[torch.rand((n_samples, 10)), torch.rand((n_samples+1, 10))], y=y,
-                   observed_mod_indicator=observed_mod_indicator, y_indicator=y_indicator)
+        MUSEDataset(Xs=[torch.rand((n_samples, 10)), torch.rand((n_samples+1, 10))], y=y)
     with pytest.raises(ValueError, match="Invalid y."):
-        MUSEDataset(Xs=Xs, y=None, observed_mod_indicator=observed_mod_indicator, y_indicator=y_indicator)
+        MUSEDataset(Xs=Xs, y=None)
     with pytest.raises(ValueError, match="Invalid y."):
-        MUSEDataset(Xs=Xs, y=torch.randint(0, 2, (n_samples+1,)),
-                   observed_mod_indicator=observed_mod_indicator, y_indicator=y_indicator)
-    with pytest.raises(ValueError, match="Invalid observed_mod_indicator."):
-        MUSEDataset(Xs=Xs, y=y, observed_mod_indicator=None, y_indicator=y_indicator)
-    with pytest.raises(ValueError, match="Invalid observed_mod_indicator."):
-        MUSEDataset(Xs=Xs, y=y, observed_mod_indicator=torch.ones((n_samples+1, 3)), y_indicator=y_indicator)
-    with pytest.raises(ValueError, match="Invalid observed_mod_indicator."):
-        MUSEDataset(Xs=Xs, y=y, observed_mod_indicator=torch.ones((n_samples, 2)), y_indicator=y_indicator)
-    with pytest.raises(ValueError, match="Invalid y_indicator."):
-        MUSEDataset(Xs=Xs, y=y, observed_mod_indicator=observed_mod_indicator, y_indicator=None)
-    with pytest.raises(ValueError, match="Invalid y_indicator."):
-        MUSEDataset(Xs=Xs, y=y, observed_mod_indicator=observed_mod_indicator,
-                   y_indicator=torch.ones((n_samples+1,)))
+        MUSEDataset(Xs=Xs, y=torch.randint(0, 2, (n_samples+1,)))
 
 
 def test_getitem(sample_data):
-    Xs, y, observed_mod_indicator, y_indicator = sample_data
-    dataset = MUSEDataset(Xs=Xs, y=y, observed_mod_indicator=observed_mod_indicator, y_indicator=y_indicator)
+    Xs, y = sample_data
+    dataset = MUSEDataset(Xs=Xs, y=y)
     for i in range(len(dataset)):
         sample = dataset[i]
         assert isinstance(sample, tuple)
@@ -98,28 +80,25 @@ def test_getitem(sample_data):
         assert y_idx.item() == y[i].item()
         observed_mod_indicator_idx = sample[2]
         assert isinstance(observed_mod_indicator_idx, torch.Tensor)
-        assert torch.all(observed_mod_indicator_idx == observed_mod_indicator[i])
         y_indicator_idx = sample[3]
         assert isinstance(y_indicator_idx, torch.Tensor)
-        assert y_indicator_idx.item() == y_indicator[i].item()
 
 
 def test_missing_values(sample_data):
-    Xs, y, observed_mod_indicator, y_indicator = sample_data
-    observed_mod_indicator[0, 0] = False
-    observed_mod_indicator[1, 1] = False
-    y_indicator[2] = False
-    dataset = MUSEDataset(Xs=Xs, y=y, observed_mod_indicator=observed_mod_indicator, y_indicator=y_indicator)
+    Xs, y = sample_data
+    Xs[0][0, :] = torch.nan
+    Xs[1][1, :] = torch.nan
+    dataset = MUSEDataset(Xs=Xs, y=y)
     sample = dataset[0]
-    assert not sample[2][0].item()
-    assert sample[2][1].item()
-    assert sample[2][2].item()
-    sample = dataset[1]
     assert sample[2][0].item()
     assert not sample[2][1].item()
-    assert sample[2][2].item()
+    assert not sample[2][2].item()
+    sample = dataset[1]
+    assert not sample[2][0].item()
+    assert sample[2][1].item()
+    assert not sample[2][2].item()
     sample = dataset[2]
-    assert not sample[3].item()
+    assert sample[3].item()
 
 
 if __name__ == "__main__":